+++ src/web/users/mod.rs
@@ -2,9 +2,9 @@ mod current_user;
 mod login;
 mod register;
 mod responses;
-mod users;
+mod update;
 
 pub use current_user::get_current_user;
 pub use login::login;
 pub use register::register;
-pub use users::update_user;
+pub use update::update_user;
